+++ Zilon.Core/Zilon.Core.Benchmark/Zilon.Core.Benchmarks.Tests.csproj
@@ -201,6 +201,10 @@
       <Project>{93E96628-DE50-42AE-A2F5-4B4B042D66E6}</Project>
       <Name>Zilon.Core</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Zilon.Emulation.Common\Zilon.Emulation.Common.csproj">
+      <Project>{36013151-86C1-431E-ABC6-AFF816D4E015}</Project>
+      <Name>Zilon.Emulation.Common</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\Microsoft.CodeAnalysis.Analyzers.2.6.1\analyzers\dotnet\cs\Microsoft.CodeAnalysis.Analyzers.dll" />
